fix(updater_test): make Stat_t-using tests cross-compile on Windows
Two tests (TestRollback_UnlinksBeforeWrite, TestReplaceBinary_UnlinksBeforeWrite)
used `syscall.Stat_t` outside the runtime t.Skip block, which broke
`GOOS=windows go test -c ./internal/updater/...`. The runtime skip prevented
the tests from running on Windows but not from being compiled.

Move both tests to a new updater_unix_test.go with a `//go:build !windows`
constraint so the package is once again cross-compilable for Windows. Behavior
is unchanged — these tests only ever ran on Unix.

Flagged by the PR #845 subagent review.
